requirements.txt
openai
groq
langchain
langchain-community
langchain-experimental
huggingface_hub
sentence-transformers
pytesseract
pillow
python-docx
python-pptx
wikipedia-api
wikipedia
colorama
python-dotenv
pydantic
graphviz
pinecone-client
serpapi
transformers
wikipedia
